+++ parsing/dialects/defaultdialect/src/main/java/org/apache/calcite/sql/defaultdialect/DefaultDialect.java
@@ -14,10 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.calcite.sql.defaultparser;
+package org.apache.calcite.sql.defaultdialect;
 
 /** Default SQL parser. */
 @SuppressWarnings("unused")
-public class Default {
+public class DefaultDialect {
   // This class is currently a place-holder to commit the src/main/java folder.
 }
