+++ parsing/dialects/defaultdialect/src/test/java/org/apache/calcite/test/DefaultDialectParserTest.java
@@ -18,14 +18,14 @@ package org.apache.calcite.test;
 
 import org.apache.calcite.sql.parser.SqlParserImplFactory;
 import org.apache.calcite.sql.parser.SqlParserTest;
-import org.apache.calcite.sql.parser.defaultparser.DefaultParserImpl;
+import org.apache.calcite.sql.parser.defaultdialect.DefaultDialectParserImpl;
 
 /**
  * Tests the "Default" SQL parser.
  */
-final class DefaultParserTest extends SqlParserTest {
+final class DefaultDialectParserTest extends SqlParserTest {
 
   @Override protected SqlParserImplFactory parserImplFactory() {
-    return DefaultParserImpl.FACTORY;
+    return DefaultDialectParserImpl.FACTORY;
   }
 }
