+++ parsing/dialects/defaultdialect/src/test/java/org/apache/calcite/test/DefaultDialectQuidemTest.java
@@ -18,7 +18,7 @@
 package org.apache.calcite.test;
 
 import org.apache.calcite.config.CalciteConnectionProperty;
-import org.apache.calcite.sql.parser.defaultparser.DefaultParserImpl;
+import org.apache.calcite.sql.parser.defaultdialect.DefaultDialectParserImpl;
 
 import net.hydromatic.quidem.Quidem;
 
@@ -26,7 +26,7 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.util.Collection;
 
-class DefaultQuidemTest extends DialectQuidemTest {
+class DefaultDialectQuidemTest extends DialectQuidemTest {
 
   static final String URL = "jdbc:calcite:";
 
@@ -39,12 +39,12 @@ class DefaultQuidemTest extends DialectQuidemTest {
    * </blockquote> */
   public static void main(String[] args) throws Exception {
     for (String arg : args) {
-      new DefaultQuidemTest().test(arg);
+      new DefaultDialectQuidemTest().test(arg);
     }
   }
 
-  DefaultQuidemTest() {
-    super(DefaultParserImpl.FACTORY);
+  DefaultDialectQuidemTest() {
+    super(DefaultDialectParserImpl.FACTORY);
   }
 
   /** For {@link QuidemTest#test(String)} parameters. */
@@ -64,7 +64,7 @@ class DefaultQuidemTest extends DialectQuidemTest {
           return DriverManager.getConnection(URL,
               CalciteAssert.propBuilder()
                   .set(CalciteConnectionProperty.PARSER_FACTORY,
-                      DefaultParserImpl.class.getName() + "#FACTORY")
+                      DefaultDialectParserImpl.class.getName() + "#FACTORY")
                   .set(CalciteConnectionProperty.MATERIALIZATIONS_ENABLED,
                       "true")
                   .set(CalciteConnectionProperty.FUN, "standard,oracle")
