+++ parsing/dialects/defaultdialect/src/test/resources/sql/dummy.iq
