+++ parsing/dialects/defaultdialect/src/test/resources/sql/join.iq
