+++ parsing/dialects/defaultdialect/src/test/resources/sql/lateral.iq
