+++ parsing/dialects/defaultdialect/src/test/resources/sql/materialized_view.iq
