+++ parsing/dialects/defaultdialect/src/test/resources/sql/scalar.iq
