+++ parsing/dialects/defaultdialect/src/test/resources/sql/spatial.iq
