+++ parsing/dialects/defaultdialect/src/test/resources/sql/stream.iq
