+++ parsing/dialects/defaultdialect/src/test/resources/sql/struct.iq
