+++ parsing/dialects/defaultdialect/src/test/resources/sql/table.iq
