+++ parsing/dialects/defaultdialect/src/test/resources/sql/table_as.iq
